% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo-package.R
\docType{package}
\name{ephylo-package}
\alias{ephylo-package}
\title{ephylo}
\description{
A package with an extended data stucture for working with phylogenies
}
\section{Functions for working with trees}{

\itemize{
  \item \code{\link{as.ephylo}}
  \item \code{\link{ephylo}}
  \item \code{\link{ephylo_preorder}}
  \item \code{\link{ephylo_postorder}}
  \item \code{\link{ephylo_tips}}
  \item \code{\link{ephylo_rclade}}
  \item \code{\link{read_newick}}
}
}

\author{
\strong{Maintainer}: Michael C. Grundler \email{mcgrundler@gmail.com}

}
