% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{ephylo_ancestors}
\alias{ephylo_ancestors}
\title{Return the ancestors of each node}
\usage{
ephylo_ancestors(phy)
}
\arguments{
\item{phy}{An object inheriting from class \code{ephylo}.}
}
\value{
A list. Each element is an integer vector of node indices 
corresponding to the ancestors of the node whose index matches its
position in the list.
}
\description{
Return the ancestors of each node
}
